Make files open in existing nvim instance when open from tmux
Created a script called tnvim (tmux nvim) that will check if we are
inside tmux, if yes then it will check for socket with tmux session name
and run nvim instance listining to it. If socket is already present, it
will use nvr to open file in the existing instance. Added alias from
nvim to call tnvim instead.